app/requirements.txt
Flask==1.1.1
Flask-WTF==0.14.2
Flask-Bootstrap4==4.0.2
flask_pymongo==2.3.0
gunicorn==20.0.4